requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
et-xmlfile==1.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
six==1.16.0
telegram==0.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
yarl==1.7.2
